accept joblistings implemented
joblistings are now acceptable (via a button in the listing page). Accepting the listing currently changes the acceptor attribute of the listing object to the email contained in the userToken cookie.

joblistings are now created with owners. On creation of a joblisting the listing's owner attribute is assigned to the email stored in the userToken cookie.

completion of joblisting actions (accepting/editing) now redirects you back to the jobslisting page.